int instead size_t for OMP constructions

+++ SF/src/SFSimulator.cpp
@@ -256,7 +256,7 @@ namespace SF
 
 #pragma omp parallel for
 
-		for (size_t i = 0; i < static_cast<size_t>(agents_.size()); ++i) 
+		for (int i = 0; i < static_cast<size_t>(agents_.size()); ++i) 
 		{
 			agents_[i]->computeNeighbors();
 			agents_[i]->computeNewVelocity();
@@ -264,7 +264,7 @@ namespace SF
 
 #pragma omp parallel for
 
-		for (size_t i = 0; i < static_cast<size_t>(agents_.size()); ++i) 
+		for (int i = 0; i < static_cast<size_t>(agents_.size()); ++i) 
 			agents_[i]->update();
     
 		globalTime_ += timeStep_;
